Added actual chunk bytes saving to file

+++ SpaceioEngine/Save/SaveData/HeaderEntry.cs
@@ -10,10 +10,10 @@ namespace ConsoleEngine.Save.SaveData
     {
         public int ChunkX { get; set; }
         public int ChunkY { get; set; }
-        public long Offset {  get; set; }
+        public int Offset {  get; set; }
         public int Length { get; set; }
 
-        public HeaderEntry(int chunkX, int chunkY, long offset, int length)
+        public HeaderEntry(int chunkX, int chunkY, int offset, int length)
         {
             ChunkX = chunkX;
             ChunkY = chunkY;
